Factor out pg part of getScore and getAvatar to helper functions
