--- routes/avatars.js
@@ -1,18 +1,0 @@
-const express = require('express');
-const router  = express.Router();
-
-module.exports = (db) => {
-  router.get("/", (req, res) => {
-    db.query(`SELECT * FROM avatars;`)
-      .then(data => {
-        const avatars = data.rows;
-        res.json({ avatars });
-      })
-      .catch(err => {
-        res
-          .status(500)
-          .json({ error: err.message });
-      });
-  });
-  return router;
-};
